jobs/bz_to_silver/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
decorator==5.2.1
fastapi==0.128.0
frozenlist==1.8.0
fsspec==2026.1.0
gcsfs==2026.1.0
google-api-core==2.29.0
google-auth==2.47.0
google-auth-oauthlib==1.2.4
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-cloud-storage-control==1.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
idna==3.11
multidict==6.7.0
numpy==2.4.1
oauthlib==3.3.1
packaging==26.0
pandas==3.0.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
yarl==1.22.0
